+++ cmd/wayfinder_cli/wayfinder_cli.go
