+++ internal/events/event_bus.go
@@ -1,4 +1,4 @@
-package core
+package events
 
 import (
 	"errors"
