+++ internal/events/eventbus_test.go
@@ -1,4 +1,4 @@
-package core
+package events
 
 import (
 	"testing"
